Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit 467424ba44ae061436de6fc46ff1a0b7d96c4f4f.

+++ libmscore/cmd.cpp
@@ -465,7 +465,7 @@ void Score::cmdAddInterval(int val, const std::vector<Note*>& nl)
                   if (v.isZero())
                         ntpc1 = ntpc2 = ntpc;
                   else {
-                        if (styleB(Sid::concertPitch)) {
+                        if (styleB(StyleIdx::concertPitch)) {
                               v.flip();
                               ntpc1 = ntpc;
                               ntpc2 = Ms::transposeTpc(ntpc, v, true);
@@ -542,7 +542,7 @@ Note* Score::setGraceNote(Chord* ch, int pitch, NoteType type, int len)
       chord->setDurationType(d);
       chord->setDuration(d.fraction());
       chord->setNoteType(type);
-      chord->setMag(ch->staff()->mag(chord->tick()) * styleD(Sid::graceNoteMag));
+      chord->setMag(ch->staff()->mag(chord->tick()) * styleD(StyleIdx::graceNoteMag));
 
       undoAddElement(chord);
       select(note, SelectType::SINGLE, 0);
@@ -709,7 +709,7 @@ Segment* Score::setNoteRest(Segment* segment, int track, NoteVal nval, Fraction 
                   // extend slur
                   //
                   Chord* chord = toNote(nr)->chord();
-                  _is.slur()->undoChangeProperty(Pid::SPANNER_TICKS, chord->tick() - _is.slur()->tick());
+                  _is.slur()->undoChangeProperty(P_ID::SPANNER_TICKS, chord->tick() - _is.slur()->tick());
                   for (ScoreElement* e : _is.slur()->linkList()) {
                         Slur* slur = toSlur(e);
                         for (ScoreElement* ee : chord->linkList()) {
@@ -1051,12 +1051,7 @@ QList<Fraction> Score::splitGapToMeasureBoundaries(ChordRest* cr, Fraction gap)
       if (tuplet) {
             if (tuplet->tuplet())
                   return flist; // do no deal with nested tuplets
-            Fraction rest = tuplet->elementsDuration();
-            for (DurationElement* de : tuplet->elements()) {
-                  if (de == cr)
-                        break;
-                  rest -= de->duration();
-                  }
+            Fraction rest = Fraction::fromTicks(tuplet->tick() + tuplet->duration().ticks() - cr->segment()->tick()) * tuplet->ratio();
             if (rest < gap)
                   qDebug("does not fit in tuplet");
             else
@@ -1390,7 +1385,7 @@ void Score::upDown(bool up, UpDownMode mode)
                                     }
                                     // store the fretting change before undoChangePitch() chooses
                                     // a fretting of its own liking!
-                                    oNote->undoChangeProperty(Pid::FRET, fret);
+                                    oNote->undoChangeProperty(P_ID::FRET, fret);
                                     }
                                     break;
                               }
@@ -1468,11 +1463,11 @@ void Score::upDown(bool up, UpDownMode mode)
             else if (staff->staffType(tick)->group() == StaffGroup::TAB) {
                   bool refret = false;
                   if (oNote->string() != string) {
-                        oNote->undoChangeProperty(Pid::STRING, string);
+                        oNote->undoChangeProperty(P_ID::STRING, string);
                         refret = true;
                         }
                   if (oNote->fret() != fret) {
-                        oNote->undoChangeProperty(Pid::FRET, fret);
+                        oNote->undoChangeProperty(P_ID::FRET, fret);
                         refret = true;
                         }
                   if (refret) {
@@ -1556,7 +1551,7 @@ static void changeAccidental2(Note* n, int pitch, int tpc)
             }
       int tpc1;
       int tpc2 = n->transposeTpc(tpc);
-      if (score->styleB(Sid::concertPitch))
+      if (score->styleB(StyleIdx::concertPitch))
             tpc1 = tpc;
       else {
             tpc1 = tpc2;
@@ -1710,7 +1705,7 @@ void Score::resetUserStretch()
             return;
 
       for (Measure* m = m1; m; m = m->nextMeasureMM()) {
-            m->undoChangeProperty(Pid::USER_STRETCH, 1.0);
+            m->undoChangeProperty(P_ID::USER_STRETCH, 1.0);
             if (m == m2)
                   break;
             }
@@ -1791,7 +1786,7 @@ void Score::cmdAddStretch(qreal val)
             stretch += val;
             if (stretch < 0)
                   stretch = 0;
-            m->undoChangeProperty(Pid::USER_STRETCH, stretch);
+            m->undoChangeProperty(P_ID::USER_STRETCH, stretch);
             }
       }
 
@@ -1818,11 +1813,11 @@ void Score::cmdResetBeamMode()
                         continue;
                   if (cr->type() == ElementType::CHORD) {
                         if (cr->beamMode() != Beam::Mode::AUTO)
-                              cr->undoChangeProperty(Pid::BEAM_MODE, int(Beam::Mode::AUTO));
+                              cr->undoChangeProperty(P_ID::BEAM_MODE, int(Beam::Mode::AUTO));
                         }
                   else if (cr->type() == ElementType::REST) {
                         if (cr->beamMode() != Beam::Mode::NONE)
-                              cr->undoChangeProperty(Pid::BEAM_MODE, int(Beam::Mode::NONE));
+                              cr->undoChangeProperty(P_ID::BEAM_MODE, int(Beam::Mode::NONE));
                         }
                   }
             }
@@ -1893,7 +1888,7 @@ bool Score::processMidiInput()
                   else
                         p = staff(staffIdx)->part();
                   if (p) {
-                        if (!styleB(Sid::concertPitch)) {
+                        if (!styleB(StyleIdx::concertPitch)) {
                               ev.pitch += p->instrument(selection().tickStart())->transpose().chromatic;
                               }
                         MScore::seq->startNote(
@@ -1917,8 +1912,13 @@ bool Score::processMidiInput()
                                                 startCmd();
                                                 cmdActive = true;
                                                 }
-                                          deleteItem(n->tieBack());
-                                          deleteItem(n);
+                                          if (n && n->tieBack()) {
+                                                Note* fn =  n->tieBack()->startNote();
+                                                deleteItem(n->tieBack());
+                                                deleteItem(n);
+                                                select(fn);
+                                                _is.moveToNextInputPos();
+                                                }
                                           }
                                     }
                               }
@@ -1928,10 +1928,13 @@ bool Score::processMidiInput()
                         startCmd();
                         cmdActive = true;
                         }
-                  if (activeMidiPitches()->empty())
-                        ev.chord = false;
-                  else
-                        ev.chord = true;
+                  if (usingNoteEntryMethod(NoteEntryMethod::REALTIME_AUTO) || usingNoteEntryMethod(NoteEntryMethod::REALTIME_MANUAL)) {
+                        // It's only a chord if multiple notes are held. (i.e. ignore shift key)
+                        if (activeMidiPitches()->empty())
+                              ev.chord = false;
+                        else
+                              ev.chord = true;
+                        }
                   // TODO: add shadow note instead of real note in realtime modes
                   // (note becomes real when realtime-advance triggered).
                   addMidiPitch(ev.pitch, ev.chord);
@@ -2227,7 +2230,7 @@ void Score::cmdMirrorNoteHead()
             if (e->type() == ElementType::NOTE) {
                   Note* note = toNote(e);
                   if (note->staff() && note->staff()->isTabStaff(note->chord()->tick()))
-                        e->undoChangeProperty(Pid::GHOST, !note->ghost());
+                        e->undoChangeProperty(P_ID::GHOST, !note->ghost());
                   else {
                         MScore::DirectionH d = note->userMirror();
                         if (d == MScore::DirectionH::AUTO)
@@ -2289,7 +2292,7 @@ void Score::cmdAddBracket()
                   }
             else if (el->type() == ElementType::ACCIDENTAL) {
                   Accidental* acc = toAccidental(el);
-                  acc->undoChangeProperty(Pid::ACCIDENTAL_BRACKET, int(AccidentalBracket::PARENTHESIS));
+                  acc->undoChangeProperty(P_ID::ACCIDENTAL_BRACKET, int(AccidentalBracket::PARENTHESIS));
                   }
             else if (el->type() == ElementType::HARMONY) {
                   Harmony* h = toHarmony(el);
@@ -2312,7 +2315,7 @@ void Score::cmdMoveRest(Rest* rest, Direction dir)
             pos.ry() -= spatium();
       else if (dir == Direction::DOWN)
             pos.ry() += spatium();
-      rest->undoChangeProperty(Pid::USER_OFF, pos);
+      rest->undoChangeProperty(P_ID::USER_OFF, pos);
       }
 
 //---------------------------------------------------------
@@ -2338,8 +2341,8 @@ void Score::cmdMoveLyrics(Lyrics* lyrics, Direction dir)
             newVerse = verse + 1;
       Lyrics* nl = cr->lyrics(newVerse, placement);
       if (nl)
-            nl->undoChangeProperty(Pid::VERSE, verse);
-      lyrics->undoChangeProperty(Pid::VERSE, newVerse);
+            nl->undoChangeProperty(P_ID::VERSE, verse);
+      lyrics->undoChangeProperty(P_ID::VERSE, newVerse);
       score()->setLayout(cr->tick());
       }
 
@@ -2842,7 +2845,7 @@ void Score::cmdResequenceRehearsalMarks()
                         if (last) {
                               QString rmText = nextRehearsalMarkText(last, rm);
                               for (ScoreElement* le : rm->linkList())
-                                    le->undoChangeProperty(Pid::TEXT, rmText);
+                                    le->undoChangeProperty(P_ID::TEXT, rmText);
                               }
                         last = rm;
                         }
@@ -2951,8 +2954,8 @@ void Score::cmdPitchUp()
       Element* el = selection().element();
       if (el && el->isLyrics())
             cmdMoveLyrics(toLyrics(el), Direction::UP);
-      else if (el && (el->isArticulation() || el->isTextBase()))
-            el->undoChangeProperty(Pid::USER_OFF, el->userOff() + QPointF(0.0, -MScore::nudgeStep * el->spatium()));
+      else if (el && (el->isArticulation() || el->isText()))
+            el->undoChangeProperty(P_ID::USER_OFF, el->userOff() + QPointF(0.0, -MScore::nudgeStep * el->spatium()));
       else if (el && el->isRest())
             cmdMoveRest(toRest(el), Direction::UP);
       else
@@ -2968,8 +2971,8 @@ void Score::cmdPitchDown()
       Element* el = selection().element();
       if (el && el->isLyrics())
             cmdMoveLyrics(toLyrics(el), Direction::DOWN);
-      else if (el && (el->isArticulation() || el->isTextBase()))
-            el->undoChangeProperty(Pid::USER_OFF, el->userOff() + QPointF(0.0, MScore::nudgeStep * el->spatium()));
+      else if (el && (el->isArticulation() || el->isText()))
+            el->undoChangeProperty(P_ID::USER_OFF, el->userOff() + QPointF(0.0, MScore::nudgeStep * el->spatium()));
       else if (el && el->isRest())
             cmdMoveRest(toRest(el), Direction::DOWN);
       else
@@ -3002,8 +3005,8 @@ void Score::cmdTimeDelete()
 void Score::cmdPitchUpOctave()
       {
       Element* el = selection().element();
-      if (el && (el->isArticulation() || el->isTextBase()))
-            el->undoChangeProperty(Pid::USER_OFF, el->userOff() + QPointF(0.0, -MScore::nudgeStep10 * el->spatium()));
+      if (el && (el->isArticulation() || el->isText()))
+            el->undoChangeProperty(P_ID::USER_OFF, el->userOff() + QPointF(0.0, -MScore::nudgeStep10 * el->spatium()));
       else
             upDown(true, UpDownMode::OCTAVE);
       }
@@ -3015,8 +3018,8 @@ void Score::cmdPitchUpOctave()
 void Score::cmdPitchDownOctave()
       {
       Element* el = selection().element();
-      if (el && (el->isArticulation() || el->isTextBase()))
-            el->undoChangeProperty(Pid::USER_OFF, el->userOff() + QPointF(0.0, MScore::nudgeStep10 * el->spatium()));
+      if (el && (el->isArticulation() || el->isText()))
+            el->undoChangeProperty(P_ID::USER_OFF, el->userOff() + QPointF(0.0, MScore::nudgeStep10 * el->spatium()));
       else
             upDown(false, UpDownMode::OCTAVE);
       }
@@ -3154,7 +3157,7 @@ void Score::cmdPadNoteDecreaseTAB()
 void Score::cmdToggleLayoutBreak(LayoutBreak::Type type)
       {
       // find measure(s)
-      QList<MeasureBase*> mbl;
+      QList<Measure*> ml;
       if (selection().isRange()) {
             Measure* startMeasure = nullptr;
             Measure* endMeasure = nullptr;
@@ -3164,68 +3167,33 @@ void Score::cmdToggleLayoutBreak(LayoutBreak::Type type)
                   return;
 #if 1
             // toggle break on the last measure of the range
-            mbl.append(endMeasure);
+            ml.append(endMeasure);
             // if more than one measure selected,
             // also toggle break *before* the range (to try to fit selection on a single line)
-            if (startMeasure != endMeasure && startMeasure->prev())
-                  mbl.append(startMeasure->prev());
+            if (startMeasure != endMeasure && startMeasure->prevMeasure())
+                  ml.append(startMeasure->prevMeasure());
 #else
             // toggle breaks throughout the selection
             for (Measure* m = startMeasure; m; m = m->nextMeasure()) {
-                  mbl.append(m);
+                  ml.append(m);
                   if (m == endMeasure)
                         break;
                   }
 #endif
             }
       else {
-            MeasureBase* mb = nullptr;
             for (Element* el : selection().elements()) {
-                  switch (el->type()) {
-                        case ElementType::HBOX:
-                        case ElementType::VBOX:
-                        case ElementType::TBOX:
-                              mb = toMeasureBase(el);
-                              break;
-                        default: {
-                              // find measure
-                              Measure* measure = toMeasure(el->findMeasure());
-                              // if measure is mmrest, then propagate to last original measure
-                              if (measure)
-                                    mb = measure->isMMRest() ? measure->mmRestLast() : measure;
-                              }
-                        }
+                  Measure* measure = toMeasure(el->findMeasure());
+                  if (measure)
+                        ml.append(measure);
                   }
-                  if (mb)
-                        mbl.append(mb);
             }
       // toggle the breaks
-      for (MeasureBase* mb: mbl) {
-            if (mb) {
-                  bool val = false;
-                  switch (type) {
-                        case LayoutBreak::Type::LINE:
-                              val = !mb->lineBreak();
-                              mb->undoSetBreak(val, type);
-                              // remove page break if appropriate
-                              if (val && mb->pageBreak())
-                                    mb->undoSetBreak(false, LayoutBreak::Type::PAGE);
-                              break;
-                        case LayoutBreak::Type::PAGE:
-                              val = !mb->pageBreak();
-                              mb->undoSetBreak(val, type);
-                              // remove line break if appropriate
-                              if (val && mb->lineBreak())
-                                    mb->undoSetBreak(false, LayoutBreak::Type::LINE);
-                              break;
-                        case LayoutBreak::Type::SECTION:
-                              val = !mb->sectionBreak();
-                              mb->undoSetBreak(val, type);
-                              break;
-                        default:
-                              break;
-                        }
-                  }
+      for (Measure* measure : ml) {
+            // if measure is mm rest, then propagate to last original measure
+            measure = measure->isMMRest() ? measure->mmRestLast() : measure;
+            if (measure)
+                  measure->undoSetBreak(!measure->lineBreak(), type);
             }
       }
 
@@ -3235,9 +3203,9 @@ void Score::cmdToggleLayoutBreak(LayoutBreak::Type type)
 
 void Score::cmdToggleMmrest()
       {
-      bool val = !styleB(Sid::createMultiMeasureRests);
+      bool val = !styleB(StyleIdx::createMultiMeasureRests);
       deselectAll();
-      undo(new ChangeStyleVal(this, Sid::createMultiMeasureRests, val));
+      undo(new ChangeStyleVal(this, StyleIdx::createMultiMeasureRests, val));
       }
 
 //---------------------------------------------------------
@@ -3246,9 +3214,9 @@ void Score::cmdToggleMmrest()
 
 void Score::cmdToggleHideEmpty()
       {
-      bool val = !styleB(Sid::hideEmptyStaves);
+      bool val = !styleB(StyleIdx::hideEmptyStaves);
       deselectAll();
-      undo(new ChangeStyleVal(this, Sid::hideEmptyStaves, val));
+      undo(new ChangeStyleVal(this, StyleIdx::hideEmptyStaves, val));
       }
 
 //---------------------------------------------------------
@@ -3258,7 +3226,7 @@ void Score::cmdToggleHideEmpty()
 void Score::cmdSetVisible()
       {
       for (Element* e : selection().elements())
-            undo(new ChangeProperty(e, Pid::VISIBLE, true));
+            undo(new ChangeProperty(e, P_ID::VISIBLE, true));
       }
 
 //---------------------------------------------------------
@@ -3268,7 +3236,7 @@ void Score::cmdSetVisible()
 void Score::cmdUnsetVisible()
       {
       for (Element* e : selection().elements())
-            undo(new ChangeProperty(e, Pid::VISIBLE, false));
+            undo(new ChangeProperty(e, P_ID::VISIBLE, false));
       }
 
 //---------------------------------------------------------
@@ -3432,7 +3400,7 @@ void Score::cmdToggleVisible()
                   continue;
             bool spannerSegment = e->isSpannerSegment();
             if (!spannerSegment || !spanners.contains(toSpannerSegment(e)->spanner()))
-                  e->undoChangeProperty(Pid::VISIBLE, !e->getProperty(Pid::VISIBLE).toBool());
+                  e->undoChangeProperty(P_ID::VISIBLE, !e->getProperty(P_ID::VISIBLE).toBool());
             if (spannerSegment)
                   spanners.insert(toSpannerSegment(e)->spanner());
             }
